% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{hashRelations}
\alias{hashRelations}
\title{construct hash table with relations}
\usage{
hashRelations(relations, idx.col = "id.x", idy.col = "id.y",
  rel.col = "relation_type")
}
\arguments{
\item{relations}{data.frame with relations and their mapping identifiers}

\item{idx.col}{columname conaining mapping identifiers}

\item{idy.col}{columname conaining mapping identifiers}

\item{rel.col}{columname containing the relations}
}
\value{
envirnoment
}
\description{
construct hash table with relations
}
\details{
constructs hash table with relations for efficient lookup
}
\author{
mvaniterson
}
