% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandaltmanplot.R
\name{icc}
\alias{icc}
\title{intra-class correlation}
\usage{
icc(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector should be same length as x}
}
\value{
intraclass correlation
}
\description{
intra-class correlation
}
\details{
intra-class correlation
}
\examples{
x <- runif(500, 0, 100)
y <- x + runif(500, -5, 5)
plot(x, y)
icc(x, y)
}
\author{
mvaniterson
}
