% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{inferRelations}
\alias{inferRelations}
\title{predict mismatches}
\usage{
inferRelations(data, n = 100, plot.it = TRUE)
}
\arguments{
\item{data}{output from allelesharing}

\item{n}{= 100 default interpolation}

\item{plot.it}{= TRUE default plot classification graph}
}
\value{
predicted mismatches
}
\description{
predict mismatches
}
\details{
predict mismatches
}
\examples{
set.seed(12345)
beta <- matrix(runif(100*10, 0,1), nrow=100)
beta[1:5, 1:5]
colnames(beta) <- paste0("sample", 1:10)
genotype <- beta2genotype(beta)
genotype[1:5, 1:5]
data <- alleleSharing(genotype)
head(data)
inferRelations(data)
}
\author{
mvaniterson
}
