% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping.R
\name{phenotyping}
\alias{phenotyping}
\title{predict phenotypes}
\usage{
phenotyping(phenotype, features, train.frac = 2/3, methods = c("ridge",
  "elastic-net", "lasso", "gbm", "mboost", "pls"), ntop = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{phenotype}{vector with phenotype either numeric of categorical}

\item{features}{feature matrix}

\item{train.frac}{fraction of the data that is used for training}

\item{methods}{regression methods used}

\item{ntop}{number of selected top features (top means correlate with the phenotype of interest)}

\item{verbose}{defaults to FALSE}

\item{...}{optional arguments}
}
\value{
list containing predictions, validation results, train and test set identifiers and selected top features
}
\description{
build prediction models for phenotype based on given matrix of features
}
\details{
all penalized regression models e.g. n << p
}
\author{
mvaniterson
}
