% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wbcc.R
\name{prediction}
\alias{prediction}
\title{prediction}
\usage{
prediction(predictor, data, covariates, transformation = function(x) x,
  ncomp = NULL, impute = TRUE, ...)
}
\arguments{
\item{predictor}{results from train}

\item{data}{matrix of 450k beta values or RNAseq counts}

\item{covariates}{matrix of covariates to correct for in the model e.g. Age and Gender}

\item{transformation}{this will depend on the model e.g. 10^x -1 if model is log10(cellPercentages+1)~. Default is no transformation.}

\item{ncomp}{optimal number of components that will be used}

\item{impute}{TRUE}

\item{...}{not used yet}
}
\value{
predicted cell percentage
}
\description{
prediction cell percentages based on trained model
}
\details{
prediction cell percentages based on trained model
}
\author{
mvaniterson
}
