% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_wbcc.R
\name{train}
\alias{train}
\title{train}
\usage{
train(data, covariates, cellPercentages, model = formula(cellPercentages ~
  covariates + data), ncomp = 50, keep.model = FALSE, ...)
}
\arguments{
\item{data}{matrix with gene expression(counts) or DNA methylation beta-values}

\item{covariates}{matrix of covariates to correct for in the model e.g. Age and Gender}

\item{cellPercentages}{matrix of cell percentages for which the predictor will be trained}

\item{model}{formula e.g. log10(cellPercentages+1)~covariates+data}

\item{ncomp}{total number of pls components that will be tested}

\item{keep.model}{logical default FALSE do not return full pls-model}

\item{...}{additional parameters for plsr see ?plsr}
}
\value{
prediction model plsr object
}
\description{
train predictor
}
\author{
mvaniterson
}
