% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{alleleSharing}
\alias{alleleSharing}
\title{allele sharing based on ibs}
\usage{
alleleSharing(x, y = NULL, rHash, phasing = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{genotype vector or matrix}

\item{y}{genotype vector or matrix}

\item{rHash}{envirnoment containing sample relationships}

\item{phasing}{FALSE}

\item{verbose}{show progress}
}
\value{
data.frame with mean and variance ibs between all pairs
}
\description{
allele sharing based on ibs
}
\details{
calculate mean variance between to vectors/matrices genotypes
coded as 1,2,3
}
\author{
mvaniterson
}
