% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{alleleSharing}
\alias{alleleSharing}
\title{allele sharing based on ibs}
\usage{
alleleSharing(x, y = NULL, relations = NULL, idx.col = "idx",
  idy.col = "idy", rel.col = "relation_type", callRate = 0.95,
  coverageRate = 2/3, phasing = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{genotype vector or matrix}

\item{y}{genotype vector or matrix}

\item{relations}{data.frame with relations and their mapping identifiers}

\item{idx.col}{columname conaining mapping identifiers}

\item{idy.col}{columname conaining mapping identifiers}

\item{rel.col}{columname containing the relations}

\item{callRate}{default 0.95 SNPs that are called in less then the threshold are dropped}

\item{coverageRate}{default 2/3 samples with less then threshold SNPs called are set to NA}

\item{phasing}{FALSE}

\item{verbose}{show progress}
}
\value{
data.frame with mean and variance ibs between all pairs
}
\description{
allele sharing based on ibs
}
\details{
calculate mean variance between to vectors/matrices genotypes
coded as 1,2,3
}
\author{
mvaniterson
}
