% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{alleleSharing}
\alias{alleleSharing}
\title{allele sharing based on ibs}
\usage{
alleleSharing(x, y = NULL, relations = NULL, idx.col = "idx",
  idy.col = "idy", rel.col = "relation_type", callRate = 0.95,
  coverageRate = 2/3, phasing = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{genotype vector or matrix}

\item{y}{genotype vector or matrix}

\item{relations}{data.frame with relations and their mapping identifiers}

\item{idx.col}{colum name containing mapping identifiers}

\item{idy.col}{colum name containing mapping identifiers}

\item{rel.col}{colum name containing the relations}

\item{callRate}{default 0.95 SNPs that are called in less then the threshold are dropped}

\item{coverageRate}{default 2/3 samples with less then threshold SNPs called are set to NA}

\item{phasing}{FALSE}

\item{verbose}{show progress default TRUE}
}
\value{
data.frame with mean and variance ibs between all pairs
}
\description{
allele sharing based on ibs
}
\details{
calculate mean variance between to vectors/matrices genotypes
coded as 1,2,3
}
\examples{
set.seed(12345)
beta <- matrix(runif(100*10, 0,1), nrow=100)
beta[1:5, 1:5]
colnames(beta) <- paste0("sample", 1:10)
genotype <- beta2genotype(beta)
genotype[1:5, 1:5]
data <- alleleSharing(genotype)
head(data)
}
\author{
mvaniterson
}
