% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{beta2genotype}
\alias{beta2genotype}
\title{converts beta-values to genotypes (1, 2 and 3)}
\usage{
beta2genotype(betas, na.rm = TRUE, minSep = 0.25, minSize = 5,
  centers = c(0.2, 0.5, 0.8))
}
\arguments{
\item{betas}{beta matrix of probes containing SNPs}

\item{na.rm}{TRUE drop cp for which no clustering was observed}

\item{minSep}{minimal separation between clusters}

\item{minSize}{size of smallest cluster (in percentage)}

\item{centers}{center of clusters}
}
\value{
matrix with genotypes
}
\description{
convert betas to genotypes
}
\details{
based on idea's from Leonard Schalkwyk (wateRmelon)
}
\examples{
set.seed(12345)
beta <- matrix(runif(100*10, 0,1), nrow=100)
beta[1:5, 1:5]
genotype <- beta2genotype(beta)
genotype[1:5, 1:5]
}
\author{
mvaniterson
}
