% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping.R
\name{inferRelations}
\alias{inferRelations}
\title{predict mismatches}
\usage{
inferRelations(data, n = 100, plot.it = TRUE)
}
\arguments{
\item{data}{output from allelesharing}

\item{n}{= 100 default interpolation}

\item{plot.it}{= TRUE default plot classification graph}
}
\value{
predicted mismatches
}
\description{
predict mismatches
}
\details{
predict mismatches
}
\author{
mvaniterson
}
